django_hw/hw_project/requirements.txt
alabaster==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
babel==2.17.0 ; python_version >= "3.12" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
django-environ==0.12.0 ; python_version >= "3.12" and python_version < "4"
django==5.1.4 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.12" and python_version < "4.0"
docutils==0.21.2 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
imagesize==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.5 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
pillow==11.1.0 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.19.1 ; python_version >= "3.12" and python_version < "4.0"
pymongo==4.10.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
snowballstemmer==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
soupsieve==2.6 ; python_version >= "3.12" and python_version < "4.0"
sphinx==8.1.3 ; python_version >= "3.12" and python_version < "4.0"
sphinxcontrib-applehelp==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
sphinxcontrib-devhelp==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
sphinxcontrib-htmlhelp==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
sphinxcontrib-jsmath==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
sphinxcontrib-qthelp==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
sphinxcontrib-serializinghtml==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
